mcp-server/requirements.txt
mcp[cli]>=0.9.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
